Added an application sandbox
The application sandbox is a wrapper used to secure internal Rubikon logic
from access by user generated application code. This is mostly to prevent
accidental execution or change of Rubikon's internal code. But it also helps
to prevent possible security problems depending on the code used inside the
application logic.